requirements.txt
google-adk
litellm
agentspeak
chromadb
pypdf
python-dotenv
typing-extensions
